correct errors in adb.py
